move import_vcf to testutils
