+++ hail/src/test/scala/is/hail/expr/ir/IRSuite.scala
@@ -747,7 +747,7 @@ class IRSuite extends SparkSuite {
     val table = TableRange(100, 10)
 
     val mt = MatrixTable.range(hc, 20, 2, Some(3)).ast.asInstanceOf[MatrixRead]
-    val vcf = hc.importVCF("src/test/resources/sample.vcf")
+    val vcf = is.hail.TestUtils.importVCF("src/test/resources/sample.vcf")
       .ast.asInstanceOf[MatrixRead]
     val bgen = hc.importBgens(FastIndexedSeq("src/test/resources/example.8bits.bgen"))
       .ast.asInstanceOf[MatrixRead]
@@ -884,7 +884,7 @@ class IRSuite extends SparkSuite {
         .ast.asInstanceOf[MatrixRead]
       val range = MatrixTable.range(hc, 3, 7, None)
         .ast.asInstanceOf[MatrixRead]
-      val vcf = hc.importVCF("src/test/resources/sample.vcf")
+      val vcf = is.hail.TestUtils.importVCF("src/test/resources/sample.vcf")
         .ast.asInstanceOf[MatrixRead]
       val bgen = hc.importBgens(FastIndexedSeq("src/test/resources/example.8bits.bgen"))
         .ast.asInstanceOf[MatrixRead]
