+++ hail/src/test/scala/is/hail/io/ImportVCFSuite.scala
@@ -1,7 +1,7 @@
 package is.hail.io
 
 import is.hail.check.Prop._
-import is.hail.SparkSuite
+import is.hail.{SparkSuite, TestUtils}
 import is.hail.annotations.Annotation
 import is.hail.io.vcf.{ExportVCF, LoadVCF}
 import is.hail.variant._
@@ -18,13 +18,13 @@ class ImportVCFSuite extends SparkSuite {
       val truth = {
         val f = tmpDir.createTempFile(extension="vcf")
         ExportVCF(vds, f)
-        hc.importVCF(f, rg = Some(vds.referenceGenome))
+        TestUtils.importVCF(f, rg = Some(vds.referenceGenome))
       }
 
       val actual = {
         val f = tmpDir.createTempFile(extension="vcf")
         ExportVCF(truth, f)
-        hc.importVCF(f, rg = Some(vds.referenceGenome))
+        TestUtils.importVCF(f, rg = Some(vds.referenceGenome))
       }
 
       truth.same(actual)
